@@ -40,14 +40,14 @@
     "@itwin/webgl-compatibility": "^4.4.0",
     "@types/history": "^4.7.11",
     "@types/node": "^18.18.10",
-    "@types/react": "^17.0.19",
-    "@types/react-dom": "^17.0.9",
+    "@types/react": "^18.2.57",
+    "@types/react-dom": "^18.2.19",
     "cross-env": "^7.0.3",
     "dotenv": "^16.3.1",
     "history": "^4.10.1",
     "npm-run-all": "^4.1.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "^7.2.9",
     "react-router-dom": "^6.10.0",
     "redux": "^4.2.1",
